Split up common.h headers
